requirements.txt
mysql-connector-python==9.6.0
python-dotenv==1.2.2
pytz==2026.1.post1
